fixed a few minor bugs and addressed some review comments.

+++ atc/db/migration/migrations/1627322254_add_last_check_id_and_plan_to_resource_config_scopes.up.sql
@@ -1,5 +1,5 @@
 ALTER TABLE resource_config_scopes
-    ADD COLUMN last_check_build_id int8,
+    ADD COLUMN last_check_build_id bigint,
     ADD COLUMN last_check_build_plan json NULL DEFAULT '{}'::json;
 
 -- This index should not be unique, because a build may check multiple resources and resource types.
